+++ src/Exceptions/DefinitionException.php
@@ -5,6 +5,12 @@ namespace Nuwave\Lighthouse\Exceptions;
 use Exception;
 use GraphQL\Error\ClientAware;
 
+/**
+ * Thrown on errors in the schema definition.
+ *
+ * This signals a developer error, so we do not
+ * show this exception to the user.
+ */
 class DefinitionException extends Exception implements ClientAware
 {
     /**